스트레스 테스트 도입하고 서버 id관리 하는부분 수정
서버 세션 풀 도입 및 스트레스 테스트 봇 추가

세션 객체 풀(ThreadSafeStack) 기반 관리로 메모리 최적화 및 동시 접속 성능 개선.
ID 타입을 int64_t로 통일하고 Actor/Player/NPC 통합 관리 방식 적용.
방 입장 자동 배정, 패킷 핸들러 개선, BehaviorTree 타입 변환 유연성 강화.
IOCP 기반 스트레스 테스트용 봇 클라이언트 프로젝트 신규 추가(400명 동시 시뮬레이션).
테스트 계획서 및 문서 추가, 코드 리팩토링 및 안정성 향상.

+++ Common/Packet.h
@@ -292,7 +292,7 @@ namespace common::packet
 	// ------------------------------------------- 채팅 관련 패킷 ------------------------------------------ //
 	struct SC_PACKET_CHAT_IN_ROOM : PacketHeader
 	{
-		long long _sender_id;
+		int64_t _sender_id;
 		uint16_t _message_length;
 	};
 	// ------------------------------------------- 디버깅용 패킷 ------------------------------------------ //
